UPD: clean conformity scores

+++ mapie/conformity_scores/conformity_scores.py
@@ -66,6 +66,7 @@ class ConformityScore(metaclass=ABCMeta):
         ----------
         y: NDArray
             Observed values.
+
         y_pred: NDArray
             Predicted values.
 
@@ -219,8 +220,8 @@ class ConformityScore(metaclass=ABCMeta):
             conformity_scores = -np.array(conformity_scores)
         y_distribution = self.get_estimation_distribution(
             y_pred, conformity_scores
-        )
+        ).astype(float)
         y_alpha = np_nanquantile(
-            y_distribution.astype(float), alpha, axis=1, method=method
+            y_distribution, alpha, axis=1, method=method
         )
         return y_alpha
